requirement.txt
Django==5.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
Pillow==9.3.0
django-cors-headers==4.0.0
